Split protocols for easier management
